fix: middleware выделить в отдельный пакет обертку врайтера
