+++ lib/intercept/docs.go
@@ -1,0 +1,4 @@
+// Содержит разные полезные утилиты для работы с хендлерами
+// и мидлварями, например оборачивался для ResponseWriter,
+// или для повторного использования (Request).body
+package intercept
